:hammer: Updated and refactored OpenMP
* This 9d07e8093a39c3b14bfc307f8afd4b84b7e57acc is most likely not true. In anycase using Rcpp::CppFunction inline has a completely different behaviour than from when the package is compile and installed. Instead of using that for the checking a utility function has been created.
* The OpenMP and Presorts have been split into two files for code-clarity.
